### 2018-03-18 #620 @leeper
@ellessenne Thanks for your submission to JOSS! Do you have anyone you'd like to recommend as a possible reviewer of the package?
