### 2018-03-18 #620 @pragyansmita
@leeper @ellessenne I am not as familiar with comorbidity codes. I can review w.r.t. the R package.
