### 2018-03-19 #620 @ellessenne
Thank you @leeper!
I would be glad if @pragyansmita could review the R package part.
Also, may I suggest @maelle? If I am correct she has experience with both R packages and epidemiology. @maelle, I would really appreciate it.
Thanks!
