### 2018-03-20 #620 @leeper
Hi @Paula-Moraga and @brooke-warson, would either of you be interested in reviewing this for JOSS?
