### 2018-03-20 #620 @brooke-watson
@maelle , thanks for thinking of me! I've never used Charlson or Elixhauser scores, so I don't think I'm the best fit given the time investment it would require. This looks like it will be a helpful package, though, @ellessenne - all the best!
