### 2018-03-20 #620 @Paula-Moraga
@leeper Same here. I have never used comorbidity scores. Sorry and thanks for thinking of me.
