### 2018-03-26 #648 @whedon
[REVIEW]: comorbidity: An R package for computing comorbidity scores
**Submitting author:** @ellessenne (<a href="http://orcid.org/0000-0002-8319-7624">Alessandro Gasparini</a>)
**Repository:** <a href="https://github.com/ellessenne/comorbidity" target ="_blank">https://github.com/ellessenne/comorbidity</a>
**Version:** 0.1.0
