### 2018-03-27 ellessenne/comorbidity#5 @corinne-riddell
Missing variables from the return descriptions for comorbidity()
`wscore` and `windex` need to be documented as part of the included dataset returned for the Charlson score (to be included ~lines 37 of [comorbidity.R](https://github.com/ellessenne/comorbidity/blob/master/R/comorbidity.R)).

Similarly for Elixhauser score, `score`, `index`, `wscore`, and `windex` should be added (~line 72 of [comorbidity.R](https://github.com/ellessenne/comorbidity/blob/master/R/comorbidity.R))

Review issue URL: https://github.com/openjournals/joss-reviews/issues/648
