### 2018-03-27 1821650
tiny edit
typo: "result being" --> "are"
