+++ inst/JOSS/paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 # Summary
 
-Comorbidity scores are extensively used in observational medical research studies to avoid potential bias when the burden of disease could be confounding the association of interest. This is of primary importance, given the increasing availability and use of administrative data (such as medical records and insurance claims) for research purposes. Several comorbidity scores have been proposed throughout the years; however, two of them result being used most frequently in practice: the Charlson comorbidity index [@Charlson_1987] and the Elixhauser comorbidity index [@Elixhauser_1998]. The Charlson comorbidity index defines a set of comorbid conditions using International Classification of Disease (ICD) diagnostic codes. Each comorbid condition has an associated weight, and the sum of all weights results in a single comorbidity score per patient. The current version of the Charlson score includes 17 comorbidities. Similarly, the Elixhauser comorbidity index is based on ICD diagnostic codes and includes 31 comorbidities. In origin, the Elixhauser index score was based on the cumulative number of conditions present; since then, several weighting systems accounting for the increase or decrease in mortality risk associated with each condition have been proposed and used in practice [@vanWalraven_2009; @Moore_2017].
+Comorbidity scores are extensively used in observational medical research studies to avoid potential bias when the burden of disease could be confounding the association of interest. This is of primary importance, given the increasing availability and use of administrative data (such as medical records and insurance claims) for research purposes. Several comorbidity scores have been proposed throughout the years; however, two of them are used most frequently in practice: the Charlson comorbidity index [@Charlson_1987] and the Elixhauser comorbidity index [@Elixhauser_1998]. The Charlson comorbidity index defines a set of comorbid conditions using International Classification of Disease (ICD) diagnostic codes. Each comorbid condition has an associated weight, and the sum of all weights results in a single comorbidity score per patient. The current version of the Charlson score includes 17 comorbidities. Similarly, the Elixhauser comorbidity index is based on ICD diagnostic codes and includes 31 comorbidities. In origin, the Elixhauser index score was based on the cumulative number of conditions present; since then, several weighting systems accounting for the increase or decrease in mortality risk associated with each condition have been proposed and used in practice [@vanWalraven_2009; @Moore_2017].
 
 `comorbidity` is an R package that allows computing comorbidity scores in an easy and straightforward way. `comorbidity` is available from the Comprehensive R Archive Network (CRAN) and from GitHub at https://github.com/ellessenne/comorbidity. The main function of the package is named `comorbidity` as the package. It allows estimating any of the supported scores, which can be selected via the `score` argument. Currently, the Charlson comorbidity score and the Elixhauser comorbidity score are supported, using either the ICD-9-CM or the ICD-10 coding system. The implementation is based on the coding algorithms published by Quan _et al_. in 2005 [@Quan_2005]. The weighting system implemented for the Charlson score is based on the original paper by Charlson _et al_. [@Charlson_1987]; conversely, we implemented the weighting system proposed by Moore _et al_. for the Elixhauser score [@Moore_2017]. Given that the version of the Elixhauser score studied by Moore _et al_. only included 29 comorbid conditions, we applied weights from van Walraven _et al_. for the remaining domains [@vanWalraven_2009]. We also categorise each comorbidity score following work by Menendez _et al_. [@Menendez_2014].
 
