### 2018-03-27 ellessenne/comorbidity#7 @corinne-riddell
README suggestions
1. It would be helpful to highlight that `samp_diag` is a function of your package. Perhaps when you illustrate how you can simulate ICD codes, you can explicitly say something like "using the  `samp_diag()` function it is possible to simulate...". I think it was a little unclear because you had written "Using `comorbidity`" and I wasn't sure if you were talking about the package or the function.

2. Similarly, further down in the README when you discuss computing the Charlson and Elixhauser scores, I think you can again add something like, "Using the `comorbidity()` function, we could compute..."

3. When you simulate x for five individuals "Conversely, say we have 5 individuals with a total of 100 ICD-9 diagnostic codes:...", you follow this by calculating the scores on x9, the three individual dataset. I think you meant to use x rather than x9 in those last two examples.

Review issue URL: https://github.com/openjournals/joss-reviews/issues/648
