### 2018-03-27 ellessenne/comorbidity#8 @corinne-riddell
testthat: no tests found
I've never used `testthat` before, so I may be doing something wrong. I cloned your repo and tried running the code in testthat.R and get the error: "Error: No tests found for comorbidity". However I was able to run the tests using `test_dir("./tests/testthat/")`.

No sure if this error was the result of my misunderstanding/misuse, but wanted to note here how I enabled the tests to run.

Review Issue URL: https://github.com/openjournals/joss-reviews/issues/648
