### 2018-03-27 #648 @corinne-riddell
Hi @ellessenne,

Nice package! I've learned a lot reading your code and it was interesting to read about the comorbidity scores based on ICD codes.

I've opened three issues  with a few suggestions to improve documentation in the README, and return descriptions for the comorbidity function. I also have an issue regarding testthat, but it may be a misunderstanding on my end.

I have also left unchecked two items above to be addressed:
- please add community guidelines to your documentation
- there is also an item re: clear statement of dependencies. I didn't find any, but not sure of the protocol if your package doesn't have any.

Once these items are addressed we can loop in the editor and keep the ball rolling.
