### 2018-03-28 #648 @ellessenne
Thank you very much @corinne-riddell for your review!
I merged your pull requests and I am going to work on the issues you raised straight away. Very helpful!
Regarding the two unchecked items left to be addressed:
1- I can add a CONTRIBUTING.md file to the repository;
2- I can add explicit dependencies on the README file, it would be the same packages included in the
DESCRIPTION file.
@leeper, would this be ok?
Thanks!
