### 2018-03-28 #648 @leeper
No need to list dependencies beyond the DESCRIPTION file. I'll tick that one for you.
