### 2018-03-28 ellessenne/comorbidity#8 @ellessenne
My workflow for running tests is:
1- Opening the RStudio project relative to the package
2- Run devtools::test()
Could you the same to see if it works for you?
Thanks!
