### 2018-03-28 72c72b6
Added CONDUCT.md and CONTRIBUTING.md
