### 2018-03-28 #648 @ellessenne
Thanks @leeper; I also just added contributing guidelines and code of conduct to the repository.

@corinne-riddell, I made some changes following your PRs / Issues, could you please give it a check? Thank you!
