### 2018-03-28 ellessenne/comorbidity#8 @corinne-riddell
That worked! I was prompted to install `roxygen2` and its dependencies, and after indicating Yes to installed, the packages installed and the tests ran. Maybe update the line in testthat.R to follow this workflow and include a comment re: opening the .Rproj file?
