### 2018-03-28 ellessenne/comorbidity#8 @ellessenne
I think that the `testthat.R` file should not be touched, R CMD check is able to find and run the tests properly (see Hadley's bible on packages development: http://r-pkgs.had.co.nz/tests.html)!
