### 2018-03-28 #648 @leeper
@ellessenne Can you please generate an archive of your repository using, for example, figshare or zenodo and then include the DOI for the archive in a comment here? Once I have that, we can accept your paper in JOSS!
