### 2018-03-29 #648 @ellessenne
Thank you very much again @corinne-riddell and @leeper!
I uploaded the current version of `comorbidity` to Zenodo, the corresponding DOI is: [10.5281/zenodo.1209837](https://doi.org/10.5281/zenodo.1209837). I think the link (via [doi.org](https://doi.org)) is not active yet, I am not sure how long it will take but should be ok!
