### 2018-03-29 #648 @ellessenne
Thank you all! This was a great experience overall, and I would like to contribute to the project - @leeper, if you need reviewers for R / epidemiology / biostatistics manuscripts please don't hesitate to contact me!
